refactor utils into their own file
